Bootstrap v4
Some errors were fixed, and then I forgorted to add BonusElgible

+++ model/MOpeNode.java
@@ -6,13 +6,13 @@ import java.io.Serializable;
 public class MOpeNode implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    public int plaintext;   // real value
+    public long plaintext;   // real value
     public long opeValue;   // OPE encrypt value
     
     public MOpeNode left;
     public MOpeNode right;
 
-    public MOpeNode(int plaintext, long opeValue) {
+    public MOpeNode(long plaintext, long opeValue) {
         this.plaintext = plaintext;
         this.opeValue = opeValue;
         this.left = null;
